fix some warnings

+++ Shawn.Utils/SimpleLogHelper.cs
@@ -267,34 +267,34 @@ namespace Shawn.Utils
             return logFileName;
         }
 
-        private void CleanUpLogFiles(FileInfo fi)
+        private void CleanUpLogFiles(FileInfo? fi)
         {
+            if (fi == null) return;
             // clean history
             if (LogFileMaxHistoryDays <= 0) return;
             var di = fi.Directory;
-            var withOutExtension = fi.Name.Substring(0, fi.Name.LastIndexOf(".", StringComparison.Ordinal));
-            var fis = di.GetFiles($"*{fi.Extension}");
-            foreach (var fileInfo in fis)
-            {
-                try
+            if (di?.GetFiles($"*{fi.Extension}") is FileInfo[] fis)
+                foreach (var fileInfo in fis)
                 {
-                    var dateStr = fileInfo.Name.Replace(fileInfo.Extension, "");
-                    dateStr = dateStr.Substring(dateStr.LastIndexOf("_") + 1);
-                    if((DateTime.Now - fileInfo.LastWriteTime).Days > LogFileMaxHistoryDays)
+                    try
                     {
-                        fileInfo.Delete();
+                        var dateStr = fileInfo.Name.Replace(fileInfo.Extension, "");
+                        dateStr = dateStr.Substring(dateStr.LastIndexOf("_") + 1);
+                        if ((DateTime.Now - fileInfo.LastWriteTime).Days > LogFileMaxHistoryDays)
+                        {
+                            fileInfo.Delete();
+                        }
+                        else if (DateTime.TryParseExact(dateStr, "yyyyMMdd", null, DateTimeStyles.None, out var date)
+                                 && date < DateTime.Now.AddDays(-1 * LogFileMaxHistoryDays))
+                        {
+                            fileInfo.Delete();
+                        }
                     }
-                    else if (DateTime.TryParseExact(dateStr, "yyyyMMdd", null, DateTimeStyles.None, out var date)
-                             && date < DateTime.Now.AddDays(-1 * LogFileMaxHistoryDays))
+                    catch (Exception)
                     {
-                        fileInfo.Delete();
+                        throw;
                     }
                 }
-                catch (Exception)
-                {
-                    throw;
-                }
-            }
         }
 
         private void MoveIfLogOverSize(string logFilePath)
@@ -353,8 +353,10 @@ namespace Shawn.Utils
                     logFileName = $"{withOutExtension}_{DateTime.Now.ToString("yyyyMMdd")}{new FileInfo(logFileName).Extension}";
 
                     var fi = new FileInfo(logFileName);
-                    // craete Directory
-                    if (!fi.Directory.Exists)
+                    // create Directory
+                    if (fi.Directory == null)
+                        return;
+                    if (fi?.Directory?.Exists == false)
                         fi.Directory.Create();
 
                     // clean history
@@ -407,7 +409,7 @@ namespace Shawn.Utils
         public string GetLog(string logFilePath, int lastLineCount = 50)
         {
             if (!File.Exists(logFilePath))
-                return null;
+                return string.Empty;
 
             var lines = File.ReadAllLines(logFilePath, Encoding.UTF8);
             var logs = new List<string>();
